# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST moderna para análise de PostgreSQL com autenticação JWT",
        "title": "PG Analytics API",
        "termsOfService": "https://pganalytics.com/terms",
        "contact": {
            "name": "Suporte PG Analytics",
            "url": "https://pganalytics.com/support",
            "email": "suporte@pganalytics.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/torresglauco/pganalytics-v2/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Autentica um usuário com username/email e senha, retornando um token JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autenticação"
                ],
                "summary": "Autenticar usuário",
                "parameters": [
                    {
                        "description": "Credenciais de login",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login bem-sucedido",
                        "schema": {
                            "$ref": "#/definitions/models.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Dados inválidos",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/metrics": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna métricas de performance e uso do sistema (rota protegida)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Métricas"
                ],
                "summary": "Obter métricas do sistema",
                "responses": {
                    "200": {
                        "description": "Métricas obtidas com sucesso",
                        "schema": {
                            "$ref": "#/definitions/models.MetricsResponse"
                        }
                    },
                    "401": {
                        "description": "Token inválido ou ausente",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ErrorResponse": {
            "description": "Resposta padrão para erros",
            "type": "object",
            "properties": {
                "error": {
                    "description": "Mensagem de erro",
                    "type": "string",
                    "example": "Invalid credentials"
                }
            }
        },
        "models.LoginRequest": {
            "description": "Dados necessários para autenticação",
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "Senha do usuário",
                    "type": "string",
                    "example": "admin123"
                },
                "username": {
                    "description": "Nome de usuário ou email",
                    "type": "string",
                    "example": "admin"
                }
            }
        },
        "models.LoginResponse": {
            "description": "Resposta da autenticação com token JWT",
            "type": "object",
            "properties": {
                "expires_in": {
                    "description": "Tempo de expiração em segundos",
                    "type": "integer",
                    "example": 86400
                },
                "token": {
                    "description": "Token JWT",
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "user": {
                    "description": "Email do usuário",
                    "type": "string",
                    "example": "admin@pganalytics.com"
                }
            }
        },
        "models.MetricsResponse": {
            "description": "Métricas de performance e uso do sistema",
            "type": "object",
            "properties": {
                "environment": {
                    "description": "Ambiente",
                    "type": "string",
                    "example": "production"
                },
                "message": {
                    "description": "Mensagem",
                    "type": "string",
                    "example": "Métricas sistema"
                },
                "metrics": {
                    "description": "Dados das métricas"
                },
                "source": {
                    "description": "Fonte dos dados",
                    "type": "string",
                    "example": "api"
                },
                "success": {
                    "description": "Sucesso da operação",
                    "type": "boolean",
                    "example": true
                },
                "timestamp": {
                    "description": "Timestamp Unix",
                    "type": "integer",
                    "example": 1672531200
                },
                "user": {
                    "description": "Dados do usuário"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Digite 'Bearer ' seguido do seu token JWT",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "tags": [
        {
            "description": "Endpoints para autenticação e gestão de usuários",
            "name": "Autenticação"
        },
        {
            "description": "Endpoints para obter métricas e dados de análise",
            "name": "Métricas"
        },
        {
            "description": "Endpoints para análise avançada do PostgreSQL",
            "name": "Analytics"
        },
        {
            "description": "Endpoints de sistema e health checks",
            "name": "Sistema"
        }
    ]
}